+++ contracts/core/test/WrappedSDTokenMock.sol
@@ -47,7 +47,7 @@ contract WrappedSDTokenMock is ERC677 {
      * @param _amount amount of unwrapped tokens
      * @return amount of wrapped tokens
      */
-    function getWrappedByUnderlying(uint256 _amount) public view returns (uint) {
+    function getWrappedByUnderlying(uint256 _amount) public view returns (uint256) {
         return _amount / multiplier;
     }
 
@@ -56,7 +56,7 @@ contract WrappedSDTokenMock is ERC677 {
      * @param _amount amount of wrapped tokens
      * @return amount of unwrapped tokens
      */
-    function getUnderlyingByWrapped(uint256 _amount) public view returns (uint) {
+    function getUnderlyingByWrapped(uint256 _amount) public view returns (uint256) {
         return _amount * multiplier;
     }
